Consolidated unit formatting into util module
