+++ data/core.yaml
@@ -319,8 +319,6 @@ en:
       metric: Metric
       imperial: Imperial
       node_count: Number of nodes
-      coordinate_pair: "{latitude}, {longitude}"
-      area_pair: "{area1} ({area2})"
   geometry:
     point: point
     vertex: vertex
@@ -1068,3 +1066,5 @@ en:
     square_meters: "{quantity} m²"
     square_kilometers: "{quantity} km²"
     hectares: "{quantity} ha"
+    area_pair: "{area1} ({area2})"
+    coordinate_pair: "{latitude}, {longitude}"
