+++ modules/ui/panels/measurement.js
@@ -8,6 +8,7 @@ import {
 } from 'd3-geo';
 
 import { t } from '../../util/locale';
+import { displayArea, displayLength, displayCoordinatePair } from '../../util/units';
 import { geoExtent } from '../../geo';
 import { utilDetect } from '../../util/detect';
 
@@ -16,7 +17,6 @@ import { utilDetect } from '../../util/detect';
 export function uiPanelMeasurement(context) {
     var locale = utilDetect().locale,
         isImperial = (locale.toLowerCase() === 'en-us');
-    var OSM_PRECISION = 7;
 
 
     function radiansToMeters(r) {
@@ -49,82 +49,6 @@ export function uiPanelMeasurement(context) {
       if (feature.type === 'Polygon') {
           return feature.coordinates[0].length - 1;
       }
-    }
-
-
-    function displayLength(m) {
-        var d = m * (isImperial ? 3.28084 : 1),
-            unit;
-
-        if (isImperial) {
-            if (d >= 5280) {
-                d /= 5280;
-                unit = 'miles';
-            } else {
-                unit = 'feet';
-            }
-        } else {
-            if (d >= 1000) {
-                d /= 1000;
-                unit = 'kilometers';
-            } else {
-                unit = 'meters';
-            }
-        }
-
-        return t('units.' + unit, {
-            quantity: d.toLocaleString(locale, { maximumSignificantDigits: 4 })
-        });
-    }
-
-
-    function displayArea(m2) {
-        var d = m2 * (isImperial ? 10.7639111056 : 1),
-            d1, d2, unit1, unit2, area;
-
-        if (isImperial) {
-            if (d >= 6969600) {     // > 0.25mi² show mi²
-                d1 = d / 27878400;
-                unit1 = 'square_miles';
-            } else {
-                d1 = d;
-                unit1 = 'square_feet';
-            }
-
-            if (d > 4356 && d < 43560000) {   // 0.1 - 1000 acres
-                d2 = d / 43560;
-                unit2 = 'acres';
-            }
-
-        } else {
-            if (d >= 250000) {    // > 0.25km² show km²
-                d1 = d / 1000000;
-                unit1 = 'square_kilometers';
-            } else {
-                d1 = d;
-                unit1 = 'square_meters';
-            }
-
-            if (d > 1000 && d < 10000000) {   // 0.1 - 1000 hectares
-                d2 = d / 10000;
-                unit2 = 'hectares';
-            }
-        }
-
-        area = t('units.' + unit1, {
-            quantity: d1.toLocaleString(locale, { maximumSignificantDigits: 4 })
-        });
-
-        if (d2) {
-            return t('info_panels.measurement.area_pair', {
-                area1: area,
-                area2: t('units.' + unit2, {
-                    quantity: d2.toLocaleString(locale, { maximumSignificantDigits: 2 })
-                })
-            });
-        } else {
-            return area;
-        }
     }
 
 
@@ -161,12 +85,7 @@ export function uiPanelMeasurement(context) {
                 .append('li')
                 .text(t('info_panels.measurement.center') + ':')
                 .append('span')
-                .text(
-                    t('info_panels.measurement.coordinate_pair', {
-                        latitude: center[1].toLocaleString(locale, { maximumFractionDigits: OSM_PRECISION }),
-                        longitude: center[0].toLocaleString(locale, { maximumFractionDigits: OSM_PRECISION })
-                    })
-                );
+                .text(displayCoordinatePair(center));
             return;
         }
 
@@ -203,7 +122,7 @@ export function uiPanelMeasurement(context) {
                     .append('li')
                     .text(t('info_panels.measurement.area') + ':')
                     .append('span')
-                    .text(displayArea(area));
+                    .text(displayArea(area, isImperial));
             }
 
 
@@ -211,18 +130,13 @@ export function uiPanelMeasurement(context) {
                 .append('li')
                 .text(lengthLabel + ':')
                 .append('span')
-                .text(displayLength(length));
+                .text(displayLength(length, isImperial));
 
             list
                 .append('li')
                 .text(t('info_panels.measurement.centroid') + ':')
                 .append('span')
-                .text(
-                    t('info_panels.measurement.coordinate_pair', {
-                        latitude: centroid[1].toLocaleString(locale, { maximumFractionDigits: OSM_PRECISION }),
-                        longitude: centroid[0].toLocaleString(locale, { maximumFractionDigits: OSM_PRECISION })
-                    })
-                );
+                .text(displayCoordinatePair(centroid));
 
             var toggle  = isImperial ? 'imperial' : 'metric';
 
@@ -250,9 +164,7 @@ export function uiPanelMeasurement(context) {
                 .append('li')
                 .text(centerLabel + ':')
                 .append('span')
-                .text(
-                    center[1].toFixed(OSM_PRECISION) + ', ' + center[0].toFixed(OSM_PRECISION)
-                );
+                .text(displayCoordinatePair(center));
         }
     }
